% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threearm.R
\docType{data}
\name{threearm}
\alias{threearm}
\title{Time to first remission data}
\format{The \code{threearm} is a data frame with 664 observations of 3 variables,
and has the following columns:
\itemize{
   \item \code{time} the observed times to first remission and censoring times
   \item \code{censor} the censoring indicator
   \item \code{group} the grouping variable
}}
\usage{
threearm
}
\description{
The data frame \code{threearm} is obtained by resampling from perturbed time-to-remission 
from patients in a three-arm randomized clinical trial for the treatment of major depression. 
See \code{\link{nocrossings}}, \code{\link{ptwiseELtest}} and \code{\link{supELtest}} for the application.
}
\seealso{
\code{\link{nocrossings}}, \code{\link{ptwiseELtest}}, \code{\link{supELtest}}
}
\keyword{datasets}
